TPLink: Break out communication into new function
This conceptually simplifies the code, so that main communication
goroutine is mainly waiting and sending channel messages.

One intentional side-effect of this change is that a socket
connection is not left open - it is opened and closed for every message.

In testing, I noticed that the HS220 seemed to drop its connection
after about a day.  Until I put in proper reconnection, go ahead and
make each request its own connection.